requirements.txt
adblockparser==0.7
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
autocommand==2.2.2
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
cheroot==10.0.1
CherryPy==18.10.0
cherrypy-cors==1.7.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cron-descriptor==1.4.5
cryptography==3.4.8
cycler==0.12.1
distro==1.9.0
Django==5.1
django-celery-beat==2.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
dnspython==2.6.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
ExifRead==2.3.2
fonttools==4.53.1
frozenlist==1.4.1
future==1.0.0
gvm-tools==24.7.0
h11==0.14
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpagentparser==1.9.5
httpcore==1.*
httpx==0.27
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
ipaddr==2.2.0
ipwhois==1.1.0
jaraco.collections==5.0.1
jaraco.context==6.0.1
jaraco.functools==4.0.2
jaraco.text==4.0.0
jiter==0.5.0
kiwisolver==1.4.5
kombu==5.4.0
lxml==4.9.4
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
more-itertools==10.4.0
multidict==6.0.5
netaddr==0.10.1
networkx==2.6.3
numpy==2.0.1
nvdlib==0.7.7
openai==0.28.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
paramiko==3.4.0
phonenumbers==8.13.43
pillow==10.4.0
portend==3.2.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
publicsuffixlist==0.9.4
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygexf==0.2.2
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.1.2
PyPDF2==1.28.6
PySocks==1.7.1
pyTelegramBotAPI==4.22.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
python-gvm==24.7.0
python-nmap==0.7.1
python-pptx==0.6.23
python-telegram-bot==21.4
python-whois==0.7.3
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
reportlab==4.2.2
requests==2.32.3
rfc3986==1.5.0
seaborn==0.13.2
secure==0.3.0
selenium==4.23.1
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
telebot==0.0.5
tempora==5.7.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4
zc.lockfile==3.0.post1
